Add caching headers to HTTP response
This commit introduces two new headers in the HTTP response
within the `CompletedOrdersController.cs` file. The `Cache-Control`
header is set to allow caching for 5 minutes, and the `Vary`
header is added to indicate that the response may vary based
on the `Accept-Encoding` request header. These changes aim to
enhance performance by enabling response data caching.